Refactor itinerary utilities by moving random ETA generation and duration formatting functions to itineraryUtils
